refactor slack routes
